Expand on anti pattern explanation
* Also renamed and moved to artifacts collection
