+++ examples/workflows/artifacts/masked-parameter-anti-pattern.yaml
@@ -47,7 +47,7 @@ spec:
         sys.path.append(os.getcwd())
         import pickle
         with open('/tmp/i', 'rb') as f:
-            i = pickle.load(f)
+            i = i or pickle.load(f)
         print(i)
       command:
       - python
